Code refactor [big one]
This commits refactor the code in several ways.

The GPE.py probably does not work anymore, but the KFoldGPE now has a
beter organization and parameters to enable / disable data on demand.
