+++ Makefile
@@ -8,9 +8,9 @@ help:
 	@echo ""
 	@echo "  help          to show this message"
 	@echo "  install       install the python libraries"
+	@echo "  kfold         train the model and use kfold (k=1 person)"
 	@echo "  train         train the model"
 	@echo "  test          test the model"
-	@echo "  kfold         train and check the model using kfold (k = 1 person)"
 	@echo ""
 
 PATH_TO_MAT := /Users/felix/github/GPE/Datasets/
@@ -18,14 +18,20 @@ PATH_TO_MAT := /Users/felix/github/GPE/Datasets/
 install:
 	@pip3 install -r requirements.txt
 
+kfold:
+	@PATH_TO_MAT=$(PATH_TO_MAT) python3 Model/KFoldGPE.py \
+		--apply-data-augmentation \
+		--apply-min-max-normalization \
+		--include-non-linear-data \
+		--include-shank-angles
+#		--apply-kalman-filter \
+#		--include-thigh-angles
+
 train:
 	@PATH_TO_MAT=$(PATH_TO_MAT) python3 Model/GPE.py
 
 test:
 	@PATH_TO_MAT=$(PATH_TO_MAT) python3 Model/TestGPE.py
-
-kfold:
-	@PATH_TO_MAT=$(PATH_TO_MAT) python3 Model/KFoldGPE.py
 
 clean:
 	@rm -rf person_*.png *.pkl5
