+++ Model/KFoldGPE.py
@@ -2,118 +2,115 @@ import os
 import argparse
 import pickle
 import time
-import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
+import sys
 import xgboost as xg
 
-from mat_to_csv import MatToCsv
-from improve_data import ImproveData
-from kfold import Kfold
 from sklearn.model_selection import cross_val_score
 from sklearn.model_selection import train_test_split
 from sklearn.metrics import mean_squared_error as MSE, r2_score
-from augmentation_methods import AugmentationMethods
-from data_preprocessor import DataPreprocessor
 
-def calculate_angular_velocity(angles, delta_t):
-    return np.concatenate(([np.nan], np.diff(angles) / delta_t))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))) # Adds the lib folder to the path
 
-def calculate_integral(angles, delta_t, velocity):
-    integral_values = []
-    integral_sum = 0
+from kfold import Kfold
+from lib.augmentation_methods import AugmentationMethods
+from lib.matlab_processor import MatlabProcessor
+from lib.data.improve import ImproveData
+from lib.data.preprocessor import DataPreprocessor
+from lib.graphics import Graphics
 
-    for i in range(len(angles) - 1):
-        integral_sum += velocity * (angles[i + 1] + angles[i]) * delta_t
-        integral_values.append(integral_sum)
 
-    # Add the initial condition (0) to the beginning of the integral_values list
-    integral_values.insert(0, 0)
+#  Process the arguments
+# -----------------------------------------------------------------------------
+parser = argparse.ArgumentParser(description="Process some arguments.")
 
-    return integral_values
+parser.add_argument('--apply-data-augmentation', action='store_true', help='Apply data augmentation to the dataset')
+parser.add_argument('--apply-min-max-normalization', action='store_true', help='Apply Min-Max normalization to the dataset')
+parser.add_argument('--apply-kalman-filter', action='store_true', help='Apply Kalman filter to the dataset')
+parser.add_argument('--include-shank-angles', action='store_true', help='Include ShankAngles from the dataset')
+parser.add_argument('--include-thigh-angles', action='store_true', help='Include ThighAngles from the dataset')
+parser.add_argument('--include-non-linear-data', action='store_true', help='Include non-linear data in the dataset')
 
-def preprocess_dataset(file_name):
-    matToCsv = MatToCsv()
-    path_to_mat = os.environ['PATH_TO_MAT']
+args = parser.parse_args()
 
-    if(path_to_mat != '' and path_to_mat != None):
-        file_name = "{}/{}".format(path_to_mat, file_name)
+apply_data_augmentation = args.apply_data_augmentation
+apply_min_max_normalization = args.apply_min_max_normalization
+apply_kalman_filter = args.apply_kalman_filter
+include_shank_angles = args.include_shank_angles
+include_thigh_angles = args.include_thigh_angles
+include_non_linear_data = args.include_non_linear_data
 
-    dataset = matToCsv.dataset(file_name)
+#  Quick check
+if not include_shank_angles and not include_thigh_angles:
+    print("!! WARNING !!")
+    raise ValueError("At least one of the following flags must be True: include_shank_angles, include_thigh_angles")
 
-    # Extract name and velocity from the file name
-    name = 'Shank' if 'Shk' in file_name else 'Thigh'
-    velocity = float(file_name.split('_')[1].split('ms')[0]) / 10
+#  Print the flags that were passed
+print("apply_data_augmentation: ", apply_data_augmentation)
+print("apply_min_max_normalization: ", apply_min_max_normalization)
+print("apply_kalman_filter: ", apply_kalman_filter)
+print("include_shank_angles: ", include_shank_angles)
+print("include_thigh_angles: ", include_thigh_angles)
+print("include_non_linear_data: ", include_non_linear_data)
 
-    # Melt the dataframe
-    melted_df = dataset.melt(ignore_index=False, var_name='subjects', value_name='values')
+print("")
 
-    # Add velocity column
-    #melted_df[name + 'Velocity'] = velocity
-
-    # Reset index
-    melted_df.reset_index(inplace=True)
-
-    # Rename columns
-    melted_df.rename(columns={'index': 'gait_percentage'}, inplace=True)
-
-    # Add 1 to the 'index_original' column to start the index from 1
-    melted_df['gait_percentage'] = melted_df['gait_percentage'] + 1
-
-    delta_t = 1  # provided by team
-
-    # Calculate angular velocity for each group
-    angles = melted_df['values'].to_numpy()
-    angular_velocity = [calculate_angular_velocity(angles[i:i + 100], delta_t) for i in range(0, len(angles), 100)]
-    angular_velocity = np.concatenate(angular_velocity)
-    melted_df[name + 'AngularVelocity'] = angular_velocity
-
-    # integral
-    integral = [calculate_integral(angles[i:i + 100], delta_t, velocity) for i in range(0, len(angles), 100)]
-    melted_df[name + 'Integral'] = np.concatenate(integral)  # Updated this line
-
-    df = melted_df.rename(columns={'values': name + 'Angles'})
-    # Reorder the DataFrame columns
-    df = df[['gait_percentage', name + 'Angles', name + 'AngularVelocity']]
-
-    return df
-
+#  This is just to generate some sort of CSV
+# -----------------------------------------------------------------------------
 print("Person,Train RMSE,Test RMSE,Train R2,Test R2")
-# Process all people
 
+
+#  Process all people
+# -----------------------------------------------------------------------------
 time_start = time.time()
 rmse_test = []
+
 for person_number in range(1, 22):
+    # We have 6 datasets, 3 for the shank and 3 for the thigh (05ms, 10ms, 15ms)
+    file_names = ['ShkAngW_05ms.mat', 'ShkAngW_10ms.mat', 'ShkAngW_15ms.mat', 'ThiAngW_05ms.mat', 'ThiAngW_10ms.mat', 'ThiAngW_15ms.mat']
+    datasets = [MatlabProcessor.to_df(file_name) for file_name in file_names]
 
-    file_names = ['ShkAngW_05ms.mat', 'ShkAngW_10ms.mat','ShkAngW_15ms.mat', 'ThiAngW_05ms.mat', 'ThiAngW_10ms.mat', 'ThiAngW_15ms.mat']
-    datasets = [preprocess_dataset(file_name) for file_name in file_names]
 
-# DataFrame concatenation
-    shankDF = pd.concat(datasets[:3], axis=0)
-    thighDF = pd.concat(datasets[3:], axis=0).drop('gait_percentage', axis=1)
+    #  DataFrame concatenation depending on the flags
+    # -------------------------------------------------------------------------
+    independent_variable_columns = []
+    if include_shank_angles and include_thigh_angles:
+        independent_variable_columns = ['ShankAngles', 'ShankAngularVelocity', 'ThighAngles', 'ThighAngularVelocity']
+        shankDF = pd.concat(datasets[:3], axis=0)
+        thighDF = pd.concat(datasets[3:], axis=0).drop('gait_percentage', axis=1)
+    elif include_shank_angles:
+        independent_variable_columns = ['ShankAngles', 'ShankAngularVelocity']
+        shankDF = pd.concat(datasets[:3], axis=0)
+        thighDF = pd.DataFrame()
+    elif include_thigh_angles:
+        independent_variable_columns = ['ThighAngles', 'ThighAngularVelocity']
+        shankDF = pd.DataFrame()
+        thighDF = pd.concat(datasets[3:], axis=0)
 
     k_fold_df = pd.concat([shankDF, thighDF], axis=1)
 
-    # Add preprocessing methods to the dataset
-    # k_fold_df['ShankAngles'] = DataPreprocessor().butterworth_filter(k_fold_df['ShankAngles'])
-    # k_fold_df['ThighAngles'] = DataPreprocessor().butterworth_filter(k_fold_df['ThighAngles'])
-    k_fold_df = DataPreprocessor().kalman_filter(k_fold_df)
-    k_fold_df = DataPreprocessor().min_max_normalization(k_fold_df)
+
+    #  Add preprocessing methods to the dataset
+    # -------------------------------------------------------------------------
+    if apply_data_augmentation:
+        k_fold_df = AugmentationMethods().augment_dataset(k_fold_df, include_shank=include_shank_angles, include_thigh=include_thigh_angles)
+
+    if apply_kalman_filter:
+        k_fold_df = DataPreprocessor().kalman_filter(k_fold_df, independent_variable_columns)
+
+    if apply_min_max_normalization:
+        k_fold_df = DataPreprocessor().min_max_normalization(k_fold_df, independent_variable_columns)
 
     k_fold_df, k_fold_person_df  = Kfold().remove_k_persons(k_fold_df, person_number)
     k_fold_df = ImproveData().add_non_linear_data(k_fold_df)       # Suggested by Mahdy
-    k_fold_df = AugmentationMethods().augment_dataset(k_fold_df)   # Adds Augmentation methods to the dataset
-
     k_fold_person_df = ImproveData().add_non_linear_data(k_fold_person_df)
 
-    train_X = k_fold_df[['ShankAngles', 'ShankAngularVelocity', 'ThighAngles', 'ThighAngularVelocity']]
-    test_X = k_fold_person_df[['ShankAngles', 'ShankAngularVelocity', 'ThighAngles', 'ThighAngularVelocity']]
+    if(include_non_linear_data):
+        independent_variable_columns += ['non_linear_1', 'non_linear_2']
 
-    # The non_linear_1 and non_linear_2 columns are added by ImproveData().add_non_linear_data
-    if('non_linear_1' in k_fold_df.columns):
-        train_X = k_fold_df[['ShankAngles', 'ShankAngularVelocity', 'ThighAngles', 'ThighAngularVelocity', 'non_linear_1', 'non_linear_2']]
-        test_X = k_fold_person_df[['ShankAngles', 'ShankAngularVelocity', 'ThighAngles', 'ThighAngularVelocity', 'non_linear_1', 'non_linear_2']]
-
+    train_X = k_fold_df[independent_variable_columns]
+    test_X = k_fold_person_df[independent_variable_columns]
     train_y = k_fold_df['gait_percentage']
     test_y = k_fold_person_df['gait_percentage']
 
@@ -180,11 +177,11 @@ for person_number in range(1, 22):
     best_xgb_r = xg.XGBRegressor(objective='reg:gamma', seed=123, max_leaves=30, learning_rate=0.11, min_child_weight=0.9)
     best_xgb_r.fit(train_X, train_y)
 
-# Save the model
+    # Save the model
     with open('xgb_kfold_model.pkl5', 'wb') as f:
         pickle.dump(best_xgb_r, f)
 
-# Prediction and RMSE for train, validation, and test sets
+    # Prediction and RMSE for train, validation, and test sets
     train_pred = best_xgb_r.predict(train_X)
     train_rmse = np.sqrt(MSE(train_y, train_pred))
     train_r2 = r2_score(train_y, train_pred)
@@ -193,27 +190,15 @@ for person_number in range(1, 22):
     test_rmse = np.sqrt(MSE(test_y, test_pred))
     test_r2 = r2_score(test_y, test_pred)
 
-# Print RMSE and R2 scores
+    # Print RMSE and R2 scores
     print("{},{},{},{},{}".format(person_number, train_rmse, test_rmse, train_r2, test_r2))
 
-# Create a DataFrame with the predicted and original values for ShankAngles and ThighAngles
-    df = pd.DataFrame({'predicted_person': test_pred, 'identity': range(1, 101)})
+    Graphics.plot_prediction_vs_identity_for_person(person_number, test_pred, test_rmse)
 
-# Create a scatter plot with predicted and original values on the x-axis, and ShankAngles and ThighAngles on the y-axis
-    plt.scatter(df['identity'], df['identity'], label='Identity Function'.format(person_number), alpha=0.3, marker='o', s=40, color='green', edgecolors='k')
-    plt.scatter(df['identity'], df['predicted_person'], label='Predicted Gait % for Person {}'.format(person_number), alpha=0.7, marker='o', s=40, color='orange', edgecolors='k')
-
-    plt.ylabel('Gait %')
-    plt.xlabel('Predicted Gait %')
-    plt.title('Prediction vs Identity for Person {}'.format(person_number))
-
-# Add RMSE to the plot
-    plt.text(0.05, 0.95, f"RMSE: {round(test_rmse, 2)}", transform=plt.gca().transAxes)
     rmse_test.append(test_rmse)
 
-    plt.legend()
-    plt.savefig('person_{}.png'.format(person_number))
-    plt.clf() # Clear the plot
+
+print("")
 
 print("Time taken: {} seconds".format(time.time() - time_start))
 print("Average RMSE: ", np.mean(rmse_test))
